Move well selection actions to well-selection; remove unused well hover actions/reducers
