+++ protocol-designer/src/well-selection/selectors.js
@@ -45,12 +45,6 @@ const numWellsSelected: Selector<number> = createSelector(
   selectedWells => Object.keys(selectedWells).length
 )
 
-// TODO Ian 2018-04-24 is 'highlightedIngredients' useful?
-// Should HOVER_WELL_BEGIN / HOVER_WELL_END be removed in favor of HIGHLIGHT_WELLS?
-// or the other way around?
-
-// const getHighlightedWells = (state: BaseState) => rootSelector(state).highlightedIngredients.wells
-
 const wellSelectionModalData: Selector<*> = createSelector(
   rootSelector,
   s => s.wellSelectionModal
